% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_BUIS.R
\name{bayesRecon_BUIS}
\alias{bayesRecon_BUIS}
\title{BUIS for Probabilistic Reconciliation of forecasts via conditioning}
\usage{
bayesRecon_BUIS(models)
}
\arguments{
\item{models}{A list of fitted models to reconcile.}
}
\value{
An object of class \code{lst_bayesRecon_BUIS}.
}
\description{
Uses the Bottom-Up Importance Sampling algorithm to draw samples from the reconciled
forecast distribution, obtained via conditioning.
}
