% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_TDcond.R
\name{bayesRecon_TDcond}
\alias{bayesRecon_TDcond}
\title{Probabilistic forecast reconciliation of mixed hierarchies via top-down conditioning}
\usage{
bayesRecon_TDcond(models)
}
\arguments{
\item{models}{A list of fitted models to reconcile.}
}
\value{
An object of class \code{lst_bayesRecon_TDcond}.
}
\description{
Uses the top-down conditioning algorithm to draw samples from the reconciled
forecast distribution. Reconciliation is performed in two steps:
first, the upper base forecasts are reconciled via conditioning,
using only the hierarchical constraints between the upper variables; then,
the bottom distributions are updated via a probabilistic top-down procedure.
}
