% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_MixCond.R
\name{bayesRecon_MixCond}
\alias{bayesRecon_MixCond}
\title{Probabilistic forecast reconciliation of mixed hierarchies via conditioning}
\usage{
bayesRecon_MixCond(models)
}
\arguments{
\item{models}{A list of fitted models to reconcile.}
}
\value{
An object of class \code{lst_bayesRecon_MixCond}.
}
\description{
Uses importance sampling to draw samples from the reconciled
forecast distribution, obtained via conditioning, in the case of a mixed hierarchy.
}
