% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_MixCond.R, R/bayesRecon_TDcond.R
\name{bayesRecon_MixCond}
\alias{bayesRecon_MixCond}
\alias{bayesRecon_TDcond}
\title{Probabilistic reconciliation of mixed hierarchies}
\usage{
bayesRecon_MixCond(models, n_samples = 1000, suppress_warnings = TRUE)

bayesRecon_TDcond(models, n_samples = 1000, suppress_warnings = TRUE)
}
\arguments{
\item{models}{A list of fitted models to reconcile.}

\item{n_samples}{Number of samples to draw from the reconciled distribution.}

\item{suppress_warnings}{If \code{TRUE}, suppress warnings from reconciliation.}
}
\description{
\code{bayesRecon_MixCond} specifies Mixed-Conditioning (Mix-Cond) reconciliation for use within
\code{reconcile()}. The method uses importance sampling to draw samples from the reconciled
forecast distribution, obtained via conditioning, in the case of a mixed hierarchy.

\code{bayesRecon_TDcond} specifies Top-Down Conditioning reconciliation for use within
\code{reconcile()}. The method uses a top-down conditioning algorithm: first, upper base forecasts
are reconciled via conditioning using only the hierarchical constraints between the
upper; then, the bottom distributions are updated via a probabilistic top-down procedure.

Reconciliation is performed when \code{forecast()} is called on the resulting model.
Marginal reconciled forecasts follow a sample distribution.
}
\references{
Zambon, L., Azzimonti, D., Rubattu, N., Corani, G. (2024).
\emph{Probabilistic reconciliation of mixed-type hierarchical time series}.
Proceedings of the Fortieth Conference on Uncertainty in Artificial Intelligence,
PMLR 244:4078-4095. \url{https://proceedings.mlr.press/v244/zambon24a.html}.
}
\seealso{
\code{\link[fabletools:reconcile]{fabletools::reconcile()}}, \code{\link[fabletools:aggregate_key]{fabletools::aggregate_key()}},
\code{\link[=bayesRecon_BUIS]{bayesRecon_BUIS()}}, \code{\link[bayesRecon:reconc_MixCond]{bayesRecon::reconc_MixCond()}}, \code{\link[bayesRecon:reconc_TDcond]{bayesRecon::reconc_TDcond()}}
}
