% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_t.R
\name{bayesRecon_t}
\alias{bayesRecon_t}
\title{Student-t forecast reconciliation}
\usage{
bayesRecon_t(models)
}
\arguments{
\item{models}{A list of fitted models to reconcile.}
}
\value{
An object of class \code{lst_bayesRecon_t}.
}
\description{
Creates an object for probabilistic forecast reconciliation using
Student-t distributions. This method applies a multivariate Student-t model
to posterior estimation, accounting for heavier tails than
normal distributions.
}
\examples{
# bayesRecon_t(base_models)
}
