% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesRecon_MixCond.R
\name{forecast.lst_bayesRecon_MixCond}
\alias{forecast.lst_bayesRecon_MixCond}
\title{forecast.lst_bayesRecon_MixCond}
\usage{
\method{forecast}{lst_bayesRecon_MixCond}(
  object,
  key_data,
  point_forecast = list(.mean = mean),
  new_data = NULL,
  n_samples = 1000,
  suppress_warnings = TRUE,
  ...
)
}
\arguments{
\item{object}{TODO}

\item{key_data}{TODO}

\item{point_forecast}{TODO}

\item{new_data}{TODO}

\item{n_samples}{number of samples for output distribution}

\item{suppress_warnings}{if TRUE warnings are not returned}

\item{...}{extra parameters to be passed on.}
}
\description{
takes a list of of models and returns a list of reconciled models
}
